backend/requirements.txt
flask>=3.0,<4
flask-sqlalchemy>=3.1,<4
flask-migrate>=4.0,<5
flask-cors>=4.0,<6
flask-login>=0.6,<1
flask-wtf>=1.2,<2
flask-limiter>=3.8,<4
email-validator>=2.0,<3
python-dotenv>=1.0,<2
resend>=2.0,<3
werkzeug>=3.0,<4
pytest>=8.0,<9
pydantic>=2.0,<3
gunicorn>=22.0,<24
numpy>=2.0,<3
matplotlib>=3.8,<4
psycopg2-binary>=2.9,<3
